Added TimeGraphView
In this version a Timegraph can be drawn in a view.  This completes toe functional additions for the first release.

A number of small bugs is fiexed, mainly regaring null references and other stupid errors.
The  LoggingView screeen now works properly. You can filter and all logging events are properly displayed. There still is one issue to fix, if the logging window is not active, no logging is written.